Crate sidechain_mc_hash

Modules

mock

Structs

McHashInherentDataProvider

Inherent data provider that provides the hash of the latest stable Cardano block observed by the Partner Chain to be included in the header of the currently produced PC block or verifies the validity of a hash included in an imported blocks’ header.

McHashInherentDigest

Enums

McHashInherentError

Constants

INHERENT_IDENTIFIER

MC_HASH_DIGEST_ID

Traits

McHashDataSource

Queries about Cardano Blocks

Functions

get_inherent_digest_value_for_block

get_mc_hash_for_block